Fix conflict with branch 1.1.7

+++ exchangis-datasource/exchangis-datasource-service/src/main/java/com/webank/wedatasphere/exchangis/datasource/service/impl/DefaultDataSourceService.java
@@ -417,7 +417,7 @@ public class DefaultDataSourceService extends AbstractDataSourceService
     public Long delete(String operator,  Long id) throws ExchangisDataSourceException {
         QueryWrapper<ExchangisJobDsBind> condition = new QueryWrapper<>();
         condition.eq("source_ds_id", id).or().eq("sink_ds_id", id);
-        Integer inUseCount = this.exchangisJobDsBindMapper.selectCount(condition);
+        Long inUseCount = this.exchangisJobDsBindMapper.selectCount(condition);
         if (inUseCount > 0) {
             throw new ExchangisDataSourceException(ExchangisDataSourceExceptionCode.CLIENT_DATASOURCE_DELETE_ERROR.getCode(), "目前存在引用依赖");
         }
