+++ exchangis-job/exchangis-job-server/src/main/java/com/webank/wedatasphere/exchangis/job/server/service/impl/ExchangisJobDsBindServiceImpl.java
@@ -32,7 +32,7 @@ public class ExchangisJobDsBindServiceImpl implements ExchangisJobDsBindService 
     public boolean inUse(Long datasourceId) {
         QueryWrapper<ExchangisJobDsBind> condition = new QueryWrapper<>();
         condition.eq("source_ds_id", datasourceId).or().eq("sink_ds_id", datasourceId);
-        Integer count = Optional.ofNullable(this.dsBindMapper.selectCount(condition)).orElse(0);
+        Long count = Optional.ofNullable(this.dsBindMapper.selectCount(condition)).orElse(0L);
         return count > 0;
     }
 }
